Write down design/plan to develop the game
Wrote down a rough design of how the game will be structured, to better
understand the requirements, as well as to use it as guide a throughout
development of the game.

This design consists mainly of the classes which out program will
contain, as well as the data (attributes) and behavior (methods) these
classes will have.