Create board class, and generate boards/grids dynamically
Created the board the class, which will contain all the logic mentioned
in the game_logic.txt design file.

In the board class, wrote the logic which will be the foundation of our
class, and allow us to perform other operations and define other
behaviors desired (like the ones described in design).