Dynamically populate grid and generate the board locations
Wrote the logic to dynamically populate the grid, by going to each
square (each element od the grid array) and assigning that square's
number as the value of the square (assigning the number of that element
as the value the element will contain to its number).

And just to be clear, the "number of the element" would be the number
the element is at, if we start counting the first element(grid[0][0]) as
n1, the second (grid[0][1]), as n2, and so on.

Also wrote the logic to map a number to its equivalent
position in the grid array, eg: 1 => [0][0], 2 => [0][1].

And a minor change, changed the name of update_available_locations class,
to what should have been its original name in the first place.