Set up Game class, and define messages to display throughout the game
Created the game class which will be responsible for connecting all
existing classes to execute out main game loop.

Defined some of the messages to be displayed throughout the game, like
'new game', 'new round', etc.

Also wrote a method that allows us to clear the terminal, which will be
useful during the game to separate the rounds.