Find the lines on the board/grid
This is the first step of developing the winning logic/algorithm of the
game. We need to find the lines, straight and diagonal, where the player
would if he placed his symbol on. The 'straight lines' can be vertical
and horizontal lines; in other words, they are the rows and columns.
The diagonal lines are self-explanatory.

Since the board a square (n x n dimension), we know it will inly ever
have two mail diagonal lines; on the start from the element 1, and the
starting from element n of the first line, and both going until the
other extreme of the board.

Also, since in the winning logic, we are working directly with the
board's grid and not the board itself (because it is significantly
simpler than the alternative - working on the board itself, and finding
the lines by their location number), we find the lines by their index
on the grid. However, since, the way we interact with the board is by
the its locations (and the players' moves are also stores like that), we
then have to convert the indices of those lines to locations so we can
interact with the board using them later on. - *hence creating the
convert_index_to_location() method on the board class*.