Find winning combinations and check if players match them
This is the second and simplest part of the winning logic/algorithm.
We find (look for) the winning combinations in the board, since we
already have the ability to do that, and check the moves made by the
players (player played locations) to see if, when we remove locations
common to the played by the player location array, and a certain
combination (another array, of locations) the remainder (intersection)
is an empty array.

If the intersection is indeed an empty array, that means there aren't
any elements (locations) in that combination that aren't part of the
player played locations. In other words, there is no item (location) in
that combination which the player hasn't played yet. That means the
player has successfully filled that whole line (which is what a
combination is) - and that, in turn, means, VICTORY!!!

So if the above condition is met, it means we have a ; however, if its
never met met, with any of the combinations, we return false to the
question [is there a] winner?