Refactor Location class and turn it into module
Realized that the then Location class was more a collection of methods
than anything, and it the criteria for being a module - "a certain class
'has a' (which in this case is ability to prompt for location)" so
thought it would be better to turn into one.

Plus, later on we might want to use it for adding more functionality
related to locations, so it would make sense to turn it into a
repository of methods - a module.